[fix] delete empty snapshots

+++ tags/views.py
@@ -335,6 +335,7 @@ def delete(request,url):
         obj=db.find_one({'url':url, 'user': unicode(request.user)})
         if obj:
             for hash in obj.get('snapshot',[]):
+                if not hash.strip(): continue
                 fname="%s/snapshots/%s" % (settings.BASE_PATH, hash)
                 if os.path.exists(fname):
                     os.unlink(fname)
